+++ tests/conftest.py
@@ -42,14 +42,14 @@ load                Runs commands in script file that is encoded as either ASCII
 py                  Invoke python command, shell, or script
 pyscript            Runs a python script file inside the console
 quit                Exits this application.
-set                 Sets a settable parameter or shows current settings of parameters.
+set                 Sets a settable parameter or shows current settings of parameters
 shell               Execute a command as if at the OS prompt.
 shortcuts           Lists shortcuts (aliases) available.
 unalias             Unsets aliases
 """
 
 # Help text for the history command
-HELP_HISTORY = """usage: history [-h] [-r | -e | -s | -o FILE | -t TRANSCRIPT | -c] [arg]
+HELP_HISTORY = """Usage: history [arg] [-h] [-r | -e | -s | -o FILE | -t TRANSCRIPT | -c]
 
 View, run, edit, save, or clear previously entered commands.
 
@@ -65,9 +65,9 @@ optional arguments:
   -r, --run             run selected history items
   -e, --edit            edit and then run selected history items
   -s, --script          script format; no separation lines
-  -o FILE, --output-file FILE
+  -o, --output-file FILE
                         output commands to a script file
-  -t TRANSCRIPT, --transcript TRANSCRIPT
+  -t, --transcript TRANSCRIPT
                         output commands and results to a transcript file
   -c, --clear           clears all history
 """